requirements.txt
colorama
matplotlib
oracledb
python-dotenv